Moved LawvereTopologies inside GategoryOfDiagrams
